Use `setup_test_component_platform` helper for alarm_control_panel entity component tests instead of `hass.components`
